+++ plugins/nf-quilt/src/main/nextflow/quilt/jep/QuiltParser.groovy
@@ -143,7 +143,7 @@ class QuiltParser {
         this.catalogName = options.get(P_CAT)
         this.options = options
         this.metadata = metadata
-        // log.debug("QuiltParser[${bucket}] for ${packageName} in ${path}")
+    // log.debug("QuiltParser[${bucket}] for ${packageName} in ${path}")
     }
 
     String parsePkg(String pkg) {
@@ -281,9 +281,9 @@ class QuiltParser {
         return options?.get(key)
     }
 
-    String toPackageString() {
+    String toPackageString(boolean forKey = false) {
         String str = "${getBucket()}"
-        if (metadata) {
+        if (metadata && !forKey) {
             str += "?${unparseQuery(metadata)}"
         }
         if (packageName) {
@@ -292,7 +292,6 @@ class QuiltParser {
             if (tag) { pkg += ":$tag" }
             str += "#package=${pkg.replace('/', '%2f')}"
         }
-        // log.debug("toPackageString: ${str}")
         return str
     }